tests/wpt/mozilla/tests/mozilla/bluetooth/requestDevice/single-filter-two-services-succeeds.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.glucose_heart_rate);
    return window.navigator.bluetooth.requestDevice(
        { filters: [{services: ['glucose', 'tx_power']}]})
    .then(device => assert_equals(device.name, 'Glucose Device'))
    .catch(error => {
        assert_unreached('Promise should have resolved. ' + error);
    });
}, 'Filter with 2 services returns a matching device.');
</script>
